Added support for the office hours page
Added a function to util.js to print out the office hours. Updated the office-hours.markdown to be able to display the table of office hours. Added and excel file that was given by Gemma. Any future excel files that are provided will just need to be formatted the same way and the select box will need to be updated